Move identifier-case utils into shared package
Relocate src/plugin/src/identifier-case/identifier-case-utils.js into src/shared/identifier-case/identifier-case-utils.js so CLI and plugin code can reuse the same normalisation helpers without cross-package imports.
